Moved Footer Files to own folder
